+++ packages/tests/utils/history.server.spec.js
@@ -1,7 +1,7 @@
 /**
  * @jest-environment node
  */
-import { objectToURLSearchParams } from '@studiometa/js-toolkit/utils/history';
+import { objectToURLSearchParams } from '@studiometa/js-toolkit/utils';
 
 describe('The `objectToURLSearchParams` method', () => {
   it('should work server side', () => {